take out payment calculations out of payment.py to avoid lint error
