+++ pay-api/src/pay_api/services/payment.py
@@ -14,7 +14,6 @@
 """Service to manage Payment model related operations."""
 from __future__ import annotations
 
-from collections import OrderedDict, defaultdict
 from dataclasses import dataclass
 from datetime import datetime, timezone
 from decimal import Decimal
@@ -44,12 +43,10 @@ from pay_api.utils.enums import (
     PaymentMethod,
     PaymentStatus,
     PaymentSystem,
-    StatementTitles,
 )
 from pay_api.utils.user_context import user_context
 from pay_api.utils.util import (
     format_currency,
-    format_datetime,
     generate_receipt_number,
     generate_transaction_number,
     get_local_formatted_date,
@@ -60,6 +57,11 @@ from ..exceptions import BusinessException
 from ..utils.errors import Error
 from .code import Code as CodeService
 from .oauth_service import OAuthService
+from .payment_calculations import (
+    build_grouped_invoice_context,
+    build_statement_context,
+    build_statement_summary_context,
+)
 from .report_service import ReportRequest, ReportService
 
 
@@ -520,8 +522,7 @@ class Payment:  # pylint: disable=too-many-instance-attributes, too-many-public-
 
             invoices = results.get("items", [])
             statement_summary = report_inputs.statement_summary
-            grouped_invoices: list[dict] = Payment._build_grouped_invoice_context(invoices, statement,
-                                                                                  statement_summary, account_info)
+            grouped_invoices: list[dict] = build_grouped_invoice_context(invoices, statement, statement_summary)
 
             has_payment_instructions = any(item.get('payment_method') == 'EFT' for item in grouped_invoices)
 
@@ -530,11 +531,11 @@ class Payment:  # pylint: disable=too-many-instance-attributes, too-many-public-
                 formatted_totals[key] = format_currency(value)
 
             template_vars = {
-                "statementSummary": Payment._build_statement_summary_context(statement_summary),
+                "statementSummary": build_statement_summary_context(statement_summary),
                 "grouped_invoices": grouped_invoices,
                 "total": formatted_totals,
                 "account": account_info,
-                "statement": Payment.build_statement_context(kwargs.get("statement")),
+                "statement": build_statement_context(kwargs.get("statement")),
             }
 
             if has_payment_instructions:
@@ -550,182 +551,6 @@ class Payment:  # pylint: disable=too-many-instance-attributes, too-many-public-
             )
         )
         return report_response
-
-    @staticmethod
-    def _build_grouped_invoice_context(invoices: List[dict], statement: dict,
-                                       statement_summary: dict, account: dict) -> list[dict]:
-        """Build grouped invoice context, with fixed payment method order."""
-        grouped = defaultdict(list)
-        for inv in invoices:
-            method = inv.get("payment_method")
-            grouped[method].append(inv)
-        result = OrderedDict()
-        first_group = True
-        for method in [m.value for m in PaymentMethod.Order]:
-            if method not in grouped:
-                continue
-            items = grouped[method]
-            method_context = {
-                "total_paid": format_currency(sum(Decimal(inv.get("paid", 0)) for inv in items)),
-                "transactions": Payment._build_transaction_rows(items),
-                "is_index_0": first_group
-            }
-            if method == PaymentMethod.EFT.value:
-                method_context["amount_owing"] = format_currency(statement.get("amount_owing", 0.00))
-                if statement.get("is_interim_statement") and statement_summary:
-                    method_context["latest_payment_date"] = statement_summary.get("latestStatementPaymentDate")
-                elif not statement.get("is_interim_statement") and statement_summary:
-                    method_context["due_date"] = format_datetime(statement_summary.get("dueDate"))
-            if method == PaymentMethod.INTERNAL.value:
-                has_staff_payment = any("routing_slip" not in inv or inv["routing_slip"] is None for inv in items)
-                method_context["is_staff_payment"] = has_staff_payment
-            else:
-                has_staff_payment = False
-            summary = Payment._calculate_invoice_summaries(items, method, statement)
-            statement_header_text = StatementTitles['DEFAULT'].value
-            if method == PaymentMethod.INTERNAL.value and has_staff_payment:
-                statement_header_text = StatementTitles['INTERNAL_STAFF'].value
-            else:
-                statement_header_text = StatementTitles[method].value
-            method_context.update({
-                "paid_summary": summary["paid"],
-                "due_summary": summary["due"],
-                "totals_summary": summary["total"],
-                "statement_header_text": statement_header_text
-            })
-            result[method] = method_context
-            first_group = False
-        grouped_invoices = [
-            {"payment_method": method, **context}
-            for method, context in result.items()
-        ]
-        return grouped_invoices
-
-    @staticmethod
-    def _calculate_invoice_summaries(invoices: List[dict], payment_method: str, statement: dict) -> dict:
-        """Calculate paid, due, and totals summary for a specific payment method."""
-        paid_total, due_total, total_total = 0, 0, 0
-        for invoice in invoices:
-            if invoice.get("payment_method") != payment_method:
-                continue
-            paid = Decimal(invoice.get("paid", 0))
-            refund = Decimal(invoice.get("refund", 0))
-            total = Decimal(invoice.get("total", 0))
-            refund_date = invoice.get("refund_date")
-            paid_total += paid
-            total_total += total
-            due_amount = total
-
-            if payment_method != PaymentMethod.EFT.value:
-                due_amount -= paid
-                if paid == 0 and refund > 0:
-                    due_amount -= refund
-            else:
-                due_amount -= paid
-                if (
-                    paid == 0
-                    and refund > 0
-                    and refund_date
-                    and parser.parse(refund_date) <= parser.parse(statement.get("to_date"))
-                ):
-                    due_amount -= refund
-            due_total += due_amount
-        return {
-            "paid": format_currency(paid_total),
-            "due": format_currency(due_total),
-            "total": format_currency(total_total),
-        }
-
-    @staticmethod
-    def _build_transaction_rows(invoices: List[dict]) -> List[dict]:
-        rows = []
-        for inv in invoices:
-            product_lines = []
-            for item in inv.get("line_items", []):
-                label = "(Cancelled) " if inv.get("status_code") == "CANCELLED" else ""
-                product_lines.append(f"{label}{item.get('description', '')}")
-            detail_lines = []
-            for detail in inv.get("details", []):
-                detail_lines.append(f"{detail.get('label', '')} {detail.get('value', '')}")
-            fee = (
-                round(inv.get("total", 0) - inv.get("service_fees", 0), 2)
-                if inv.get("total") and inv.get("service_fees")
-                else 0.00
-            )
-            service_fee = round(inv.get("service_fees", 0), 2)
-            gst = round(inv.get("gst", 0), 2)
-            total = round(inv.get("total", 0), 2)
-            row = {
-                "products": product_lines,
-                "details": detail_lines,
-                "folio": inv.get("folio_number") or "-",
-                "created_on": format_datetime(datetime.fromisoformat(inv["created_on"]).strftime("%b %d,%Y")
-                                              if inv.get("created_on") else "-"),
-                "fee": format_currency(fee),
-                "service_fee": format_currency(service_fee),
-                "gst": format_currency(gst),
-                "total": format_currency(total),
-            }
-            skip_keys = {"details", "folio_number", "created_on", "fee", "gst", "total", "service_fees"}
-            row.update((k, v) for k, v in inv.items() if k not in skip_keys)
-            rows.append(row)
-
-        return rows
-
-    @staticmethod
-    def build_statement_context(statement: dict) -> dict:
-        """Build and enhance statement context with formatted fields."""
-        if not statement:
-            return statement
-
-        enhanced_statement = statement.copy()
-
-        from_date = format_datetime(statement.get('from_date'))
-        to_date = format_datetime(statement.get('to_date'))
-        created_on = format_datetime(statement.get('created_on'))
-        frequency = statement.get('frequency', '')
-
-        if frequency == 'DAILY' and from_date:
-            enhanced_statement['duration'] = from_date
-        elif from_date and to_date:
-            enhanced_statement['duration'] = f"{from_date} - {to_date}"
-        elif from_date:
-            enhanced_statement['duration'] = from_date
-
-        amount_owing = statement.get('amount_owing')
-        enhanced_statement['amount_owing'] = format_currency(amount_owing) if amount_owing else format_currency(0)
-
-        if from_date:
-            enhanced_statement['from_date'] = from_date
-        if to_date:
-            enhanced_statement['to_date'] = to_date
-        if created_on:
-            enhanced_statement['created_on'] = created_on
-
-        return enhanced_statement
-
-    @staticmethod
-    def _build_statement_summary_context(statement_summary: dict) -> dict:
-        if not statement_summary:
-            return None
-        enhanced_statement_summary = statement_summary.copy()
-        last_statement_total = statement_summary.get('lastStatementTotal')
-        last_statement_paid_amount = statement_summary.get('lastStatementPaidAmount')
-        cancelled_transactions = statement_summary.get('cancelledTransactions')
-        latest_statement_sayment_date = statement_summary.get('latestStatementPaymentDate')
-        due_date = statement_summary.get('dueDate')
-        enhanced_statement_summary['lastStatementTotal'] = format_currency(last_statement_total)
-        enhanced_statement_summary['lastStatementPaidAmount'] = format_currency(last_statement_paid_amount)
-        if cancelled_transactions not in [None, 0, '0', '0.00']:
-            enhanced_statement_summary['cancelledTransactions'] = format_currency(cancelled_transactions)
-        if not latest_statement_sayment_date:
-            enhanced_statement_summary['latestStatementPaymentDate'] = None
-        else:
-            enhanced_statement_summary['latestStatementPaymentDate'] = format_datetime(latest_statement_sayment_date,
-                                                                                       "%B %d, %Y")
-
-        enhanced_statement_summary['dueDate'] = format_datetime(due_date, "%B %d, %Y")
-        return enhanced_statement_summary
 
     @staticmethod
     def _prepare_csv_data(results):
